+++ pkg/convert/parser_test.go
@@ -15,7 +15,7 @@ var _ = Describe("convert", func() {
 		It("parses data correctly", func() {
 			r := bytes.NewReader([]byte("\x00\x00\x01\x06foo;ba\x00\x02\x01r\x01\x00\x01z\x01\x00"))
 			result := []string{}
-			ParseTrie(r, func(name []byte, val int) {
+			ParseTrie(r, func(name []byte, val uint64) {
 				result = append(result, fmt.Sprintf("%s %d", name, val))
 			})
 			Expect(result).To(ConsistOf("foo;bar 1", "foo;baz 1"))
@@ -45,7 +45,7 @@ var _ = Describe("convert", func() {
 		It("parses data correctly", func() {
 			r := bytes.NewReader([]byte("foo;bar 10\nfoo;baz 20\n"))
 			result := []string{}
-			ParseGroups(r, func(name []byte, val int) {
+			ParseGroups(r, func(name []byte, val uint64) {
 				result = append(result, fmt.Sprintf("%s %d", name, val))
 			})
 			Expect(result).To(ConsistOf("foo;bar 10", "foo;baz 20"))
@@ -56,7 +56,7 @@ var _ = Describe("convert", func() {
 		It("parses data correctly", func() {
 			r := bytes.NewReader([]byte("foo;bar\nfoo;baz\n"))
 			result := []string{}
-			ParseIndividualLines(r, func(name []byte, val int) {
+			ParseIndividualLines(r, func(name []byte, val uint64) {
 				result = append(result, fmt.Sprintf("%s %d", name, val))
 			})
 			Expect(result).To(ConsistOf("foo;bar 1", "foo;baz 1"))
